requirements.txt
alembic==1.8.1
APScheduler==3.9.1
Cartopy==0.21.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
fonttools==4.37.4
healthcheck==1.3.3
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
Mako==1.2.3
MarkupSafe==2.1.1
matplotlib==3.6.0
networkx==2.8.7
numpy==1.23.3
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
psycopg2==2.9.3
pyparsing==3.0.9
pyproj==3.4.0
pyshp==2.3.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
scipy==1.9.1
Shapely==1.8.4
six==1.16.0
SQLAlchemy==1.4.41
tropycal==0.6
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
Werkzeug==2.2.2
xarray==2022.9.0
zipp==3.8.1
